tests

+++ Sources/motor.c
@@ -73,7 +73,7 @@ uint8_t setMotorDirection(uint8_t motor, enum MotorDirection val) {
 }
 
 /* set motor speed */
-uint8_t setMotorSpeed(uint8_t motor, uint8_t val) {
+uint8_t setMotorSpeed(uint8_t motor, uint16_t val) {
 	if (motor >= 0 && motor < NUMBER_OF_MOTOR && (val >= 0 && val <= 0xFFFF)) {
 		switch (motor) {
 		case 0:
